requirements.txt
httpx~=0.21.1
pyyaml~=6.0
regex~=2021.11.10
